+++ NEWS.md
@@ -4,3 +4,4 @@
 + Added main function `findVariableFeaturesBayes()` (still in development right now). 
 + Added helper function `sampleMarginal()` to make sampling from posterior marginal distribution possible. 
 + Added support for adding gene statistics metadata to `SingleCellExperiment` or `Seurat` objects after estimation. 
++ Changed main model backend to `brms` via `cmdstanr` instead of `INLA`, since `INLA` doesn't appear to support per-group estimation of the negative-binomial dispersion parameter.